Kill the g_game warnings.  If there are problems feel free to blame me.
SVN r2057 (0.6-dev)

+++ client/src/g_game.cpp
@@ -523,7 +523,7 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	}
 
 	// Joystick analog strafing -- Hyper_Eye
-	side += (((float)joystrafe / (float)SHRT_MAX) * sidemove[speed]);
+	side += (int)(((float)joystrafe / (float)SHRT_MAX) * sidemove[speed]);
 
 	if (Actions[ACTION_LOOKUP])
 		look += lookspeed[speed];
@@ -554,9 +554,9 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	if(joy_freelook && sv_freelook)
 	{
 		if (joy_invert)
-			look += (((float)joylook / (float)SHRT_MAX) * lookspeed[speed]);
+			look += (int)(((float)joylook / (float)SHRT_MAX) * lookspeed[speed]);
 		else
-			look -= (((float)joylook / (float)SHRT_MAX) * lookspeed[speed]);
+			look -= (int)(((float)joylook / (float)SHRT_MAX) * lookspeed[speed]);
 	}
 
 	if (Actions[ACTION_MOVERIGHT])
@@ -588,20 +588,20 @@ void G_BuildTiccmd (ticcmd_t *cmd)
 	Impulse = 0;
 
 	if (strafe || lookstrafe)
-		side += (((float)joyturn / (float)SHRT_MAX) * sidemove[speed]);
+		side += (int)(((float)joyturn / (float)SHRT_MAX) * sidemove[speed]);
 	else
-		cmd->ucmd.yaw -= (((float)joyturn / (float)SHRT_MAX) * angleturn[1]) * (joy_sensitivity / 10);
+		cmd->ucmd.yaw -= (short)((((float)joyturn / (float)SHRT_MAX) * angleturn[1]) * (joy_sensitivity / 10));
 
 	if (Actions[ACTION_MLOOK])
 	{
 		if (joy_invert)
-			look += (((float)joyforward / (float)SHRT_MAX) * lookspeed[speed]);
+			look += (int)(((float)joyforward / (float)SHRT_MAX) * lookspeed[speed]);
 		else
-			look -= (((float)joyforward / (float)SHRT_MAX) * lookspeed[speed]);
+			look -= (int)(((float)joyforward / (float)SHRT_MAX) * lookspeed[speed]);
 	}
 	else
 	{
-		forward -= (((float)joyforward / (float)SHRT_MAX) * forwardmove[speed]);
+		forward -= (int)(((float)joyforward / (float)SHRT_MAX) * forwardmove[speed]);
 	}
 
 	if ((Actions[ACTION_MLOOK]) || (cl_mouselook && sv_freelook))
@@ -778,21 +778,21 @@ BOOL G_Responder (event_t *ev)
 			{
 				mousexleft = ev->data2;
 				mousexleft = -mousexleft;
-				mousex = pow((ev->data2 * (mouse_sensitivity + 5) / 10), dynresval);
+				mousex = (int) pow((ev->data2 * (mouse_sensitivity + 5) / 10), dynresval);
 
 				if (ev->data2 < 0)
 				{
-					mousexleft = pow((mousexleft * (mouse_sensitivity + 5) / 10), dynresval);
+					mousexleft = (int) pow((mousexleft * (mouse_sensitivity + 5) / 10), dynresval);
 					mousex = -mousexleft;
 				}
 
 				mouseydown = ev->data3;
 				mouseydown = -mouseydown;
-				mousey = pow((ev->data3 * (mouse_sensitivity + 5) / 10), dynresval);
+				mousey = (int) pow((ev->data3 * (mouse_sensitivity + 5) / 10), dynresval);
 
 				if (ev->data3 < 0)
 				{
-					mouseydown = pow((mouseydown * (mouse_sensitivity + 5) / 10), dynresval);
+					mouseydown = (int) pow((mouseydown * (mouse_sensitivity + 5) / 10), dynresval);
 					mousey = -mouseydown;
 				}
 			}
@@ -801,28 +801,28 @@ BOOL G_Responder (event_t *ev)
 		{
 			if (dynres_state == 0)
 			{
-				mousex = ev->data2 * (zdoomsens); // [Toke - Mouse] Zdoom mouse code
-				mousey = ev->data3 * (zdoomsens);
+				mousex = (int)(ev->data2 * (zdoomsens)); // [Toke - Mouse] Zdoom mouse code
+				mousey = (int)(ev->data3 * (zdoomsens));
 			}
 			else if (dynres_state == 1)
 			{
 				mousexleft = ev->data2;
 				mousexleft = -mousexleft;
-				mousex = pow((ev->data2 * (zdoomsens)), dynresval);
+				mousex = (int) pow((ev->data2 * (zdoomsens)), dynresval);
 
 				if (ev->data2 < 0)
 				{
-					mousexleft = pow((mousexleft * (zdoomsens)), dynresval);
+					mousexleft = (int) pow((mousexleft * (zdoomsens)), dynresval);
 					mousex = -mousexleft;
 				}
 
 				mouseydown = ev->data3;
 				mouseydown = -mouseydown;
-				mousey = pow((ev->data3 * (zdoomsens)), dynresval);
+				mousey = (int) pow((ev->data3 * (zdoomsens)), dynresval);
 
 				if (ev->data3 < 0)
 				{
-					mouseydown = pow((mouseydown * (zdoomsens)), dynresval);
+					mouseydown = (int) pow((mouseydown * (zdoomsens)), dynresval);
 					mousey = -mouseydown;
 				}
 			}
